Add a News posts page and add to About menu
Jo requested that we add a News feature to the new Cirrus website.
I have implemented this in Hugo with an index (list) page, with links to each post.
It will benefit from some review of styling and adding of additional data such as Author, but I believe it is sufficient and functional for now